+++ examples/src/integTest/scala/ai/h2o/sparkling/LocalTestSuite.scala
@@ -1,0 +1,115 @@
+/*
+* Licensed to the Apache Software Foundation (ASF) under one or more
+* contributor license agreements.  See the NOTICE file distributed with
+* this work for additional information regarding copyright ownership.
+* The ASF licenses this file to You under the Apache License, Version 2.0
+* (the "License"); you may not use this file except in compliance with
+* the License.  You may obtain a copy of the License at
+*
+*    http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+package ai.h2o.sparkling
+
+import ai.h2o.sparkling.examples._
+import org.junit.runner.RunWith
+import org.scalatest.FunSuite
+import org.scalatest.junit.JUnitRunner
+
+@RunWith(classOf[JUnitRunner])
+class LocalTestSuite extends FunSuite with IntegTestHelper {
+  test("Launch AirlinesWithWeatherDemo", LocalTest) {
+    launch(AirlinesWithWeatherDemo.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "2g")
+        conf("spark.driver.memory", "2g")
+      }
+    )
+  }
+
+  test("Launch Chicago Crime Demo", LocalTest) {
+    launch(ChicagoCrimeAppSmall.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "3g")
+        conf("spark.driver.memory", "3g")
+      }
+    )
+  }
+
+  test("Launch Craigslist App Demo", LocalTest) {
+    launch(CraigslistJobTitlesApp.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "2g")
+        conf("spark.driver.memory", "2g")
+      }
+    )
+  }
+
+  test("Launch DeepLearnigDemo", LocalTest) {
+    launch(DeepLearningDemo.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "2g")
+        conf("spark.driver.memory", "2g")
+      }
+    )
+  }
+
+  test("Launch DeepLearningDemoWithoutExtension", LocalTest) {
+    launch(DeepLearningDemoWithoutExtension.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "2g")
+        conf("spark.driver.memory", "2g")
+      }
+    )
+  }
+
+  test("Launch HamOrSpamDemo", LocalTest) {
+    launch(HamOrSpamDemo.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "2g")
+        conf("spark.driver.memory", "2g")
+      }
+    )
+  }
+
+  test("Launch ProstateDemo", LocalTest) {
+    launch(ProstateDemo.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "2g")
+        conf("spark.driver.memory", "2g")
+      }
+    )
+  }
+
+  test("Launch simple ML pipeline using H2O", LocalTest) {
+    launch(PubDev457.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "2g")
+        conf("spark.driver.memory", "2g")
+      }
+    )
+  }
+
+  test("Verify scoring on 0-length chunks", LocalTest) {
+    launch(PubDev928.getClass.getName.replace("$", ""),
+      env {
+        sparkMaster("local[*]")
+        conf("spark.executor.memory", "2g")
+        conf("spark.driver.memory", "2g")
+      }
+    )
+  }
+}
