+++ examples/src/main/scala/ai/h2o/sparkling/examples/PubDev928.scala
@@ -1,0 +1,71 @@
+package ai.h2o.sparkling.examples
+
+import java.io.File
+
+import hex.deeplearning.DeepLearning
+import hex.deeplearning.DeepLearningModel.DeepLearningParameters
+import org.apache.spark.SparkContext
+import org.apache.spark.h2o.H2OContext
+import org.apache.spark.rdd.RDD
+import org.apache.spark.sql.SparkSession
+import water.fvec.H2OFrame
+import water.support.{H2OFrameSupport, SparkContextSupport}
+
+object PubDev928 extends SparkContextSupport with IntegTestStopper {
+
+  def main(args: Array[String]): Unit = exitOnException {
+    val conf = configure("PUBDEV-928")
+    val sc = new SparkContext(conf)
+    val h2oContext = H2OContext.getOrCreate()
+    implicit val sqlContext = SparkSession.builder().getOrCreate().sqlContext
+    import sqlContext.implicits._
+
+    val airlinesData = new H2OFrame(new File(TestUtils.locate("smalldata/airlines/allyears2k_headers.zip")))
+
+    // We need to explicitly repartition data to 12 chunks/partitions since H2O parser handling
+    // partitioning dynamically based on available number of CPUs
+    println("Number of chunks before query: " + airlinesData.anyVec().nChunks())
+    val airlinesTable: RDD[Airlines] = h2oContext.asRDD[Airlines](airlinesData)
+    airlinesTable.toDF.createOrReplaceTempView("airlinesTable")
+
+    val query = "SELECT * FROM airlinesTable WHERE Dest LIKE 'SFO'"
+    // Transform result of SQL query directly into H2OFrame, but change number of
+    val queryResult = sqlContext.sql(query)
+    val partitionNumbers = queryResult.count().asInstanceOf[Int] + 1
+    val result: H2OFrame = h2oContext.asH2OFrame(queryResult.repartition(partitionNumbers), "flightTable")
+    println("Number of partitions in query result: " + queryResult.rdd.partitions.length)
+    println("Number of chunks in query result" + result.anyVec().nChunks())
+
+    val train: H2OFrame = result('Year, 'Month, 'DayofMonth, 'DayOfWeek, 'CRSDepTime, 'CRSArrTime,
+      'UniqueCarrier, 'FlightNum, 'TailNum, 'CRSElapsedTime, 'Origin, 'Dest,
+      'Distance, 'IsDepDelayed)
+    H2OFrameSupport.withLockAndUpdate(train) { fr =>
+      fr.replace(fr.numCols() - 1, fr.lastVec().toCategoricalVec)
+    }
+    println(s"Any vec chunk cnt: ${train.anyVec().nChunks()}")
+    // Configure Deep Learning algorithm
+    val dlParams = new DeepLearningParameters()
+    dlParams._train = train.key
+    dlParams._response_column = "IsDepDelayed"
+
+    val dl = new DeepLearning(dlParams)
+    val dlModel = dl.trainModel.get
+
+    // THIS WILL FAIL
+    val testFrame: H2OFrame = result
+    // Verify that testFrame has at least one chunk with 0-rows
+    val av = testFrame.anyVec()
+    println(s"Test frame chunk cnt: ${av.nChunks()}")
+    for (i <- 0 until av.nChunks()) {
+      println(av.chunkForChunkIdx(i).len())
+    }
+
+    // And run scoring on dataset which contains at least one chunk with zero-lines
+    import h2oContext.implicits._
+    val predictionH2OFrame = dlModel.score(testFrame)('predict)
+    assert(predictionH2OFrame.numRows() == testFrame.numRows())
+
+    // Shutdown Spark cluster and H2O
+    h2oContext.stop(stopSparkContext = true)
+  }
+}
